fix calculation bug

+++ index.js
@@ -95,7 +95,7 @@ app.post('/hasil/nelayan', function(request, response) {
                     console.log('----------------------')
         
                     var berat_sekarang = model.get('berat_total')
-                    berat_sekarang += parseInt(berat)
+                    berat_sekarang += parseFloat(berat)
         
                     // Update hasil_ikan berat
                     new hasil_ikan.model({
